Backend/app/requirements.txt
absl-py==2.3.1
accelerate==1.9.0
ag2==0.9.9
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.9
autogen-agentchat==0.7.4
autogen-core==0.7.4
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.3
courlan==1.3.2
cryptography==45.0.5
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.2
deep-translator==1.11.4
deepdiff==6.7.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docx2txt==0.9
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.0.32
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langdetect==1.0.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
lxml==5.4.0
lxml_html_clean==0.4.2
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mesop==1.1.0
mpmath==1.3.0
msgpack==1.1.1
msoffcrypto-tool==5.4.2
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
openai==1.101.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.36.0
ordered-set==4.1.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==23.2
pandas==2.3.1
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.0.0
pikepdf==9.10.2
pillow==11.3.0
pinecone-client==3.2.2
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.10.0
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
streamlit==1.47.1
sympy==1.14.0
tenacity==8.5.0
termcolor==3.1.0
tiktoken==0.9.0
timm==1.0.19
tld==0.13.1
tokenizers==0.21.4
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.54.1
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
unstructured==0.18.11
unstructured-client==0.41.0
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
